refactor: Rename files to more logic variant without redundancy
