+++ src/components/App.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Header from './header/SearchBar';
+import Header from './Header/SearchBar';
 import Profile from './profile/Profile';
 import './App.css';
